+++ tests/test_base.py
@@ -5,14 +5,15 @@ from http import HTTPStatus
 # noinspection PyPackageRequirements
 from werkzeug.wrappers import Response
 
-from app import create_test_app
-from tests.utils import TestJwt
+from flask import Response
+
+from tests.utils import TestJwt, create_app
 from flask_azure_oauth.errors import ApiAuthError
 
 
 class FlaskOAuthProviderBaseTestCase(unittest.TestCase):
     def setUp(self):
-        self.app = create_test_app()
+        self.app = create_app()
         self.app_context = self.app.app_context()
         self.app_context.push()
         self.client = self.app.test_client()
@@ -61,13 +62,10 @@ class FlaskOAuthProviderBaseTestCase(unittest.TestCase):
         )
         return jwt.dumps()
 
-    def _change_application_auth(self, mode: str = None):
+    def _change_application_auth(self, **kwargs):
         self.app_context.pop()
 
-        if mode is not None:
-            self.app = create_test_app(AUTH_MODE=mode)
-        else:
-            self.app = create_test_app()
+        self.app = create_app(**kwargs)
 
         self.app_context = self.app.app_context()
         self.app_context.push()