+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 * Support for access tokens from the Flask session (using `access_token` as a key)
+* Minimal support for custom signing keys (app specific JWKS)
 
 ### Fixed